+++ cineflow/bases/singleton.py
