--- requirements.txt
@@ -1,3 +1,0 @@
-pyyaml>=6.0
-requests>=2.32.3
-pillow>=11.2.1
